Move getCluter selector with other cluster selectors
