+++ assets/js/components/ClusterDetails/ClusterDetails.jsx
@@ -20,8 +20,7 @@ import ProviderLabel from '@components/ProviderLabel';
 import { getClusterName } from '@components/ClusterLink';
 import { EOS_SETTINGS, EOS_CLEAR_ALL, EOS_PLAY_CIRCLE } from 'eos-icons-react';
 
-import { getCluster } from '@state/selectors';
-import { getClusterHostIDs } from '@state/selectors/cluster';
+import { getCluster, getClusterHostIDs } from '@state/selectors/cluster';
 import {
   updateLastExecution,
   executionRequested,
