+++ assets/js/components/ClusterDetails/ClusterSettings.jsx
@@ -12,8 +12,7 @@ import WarningBanner from '@components/Banners/WarningBanner';
 import { getClusterName } from '@components/ClusterLink';
 import { ClusterInfoBox } from '@components/ClusterDetails';
 
-import { getCluster } from '@state/selectors';
-import { getClusterHostIDs } from '@state/selectors/cluster';
+import { getCluster, getClusterHostIDs } from '@state/selectors/cluster';
 
 export const UNKNOWN_PROVIDER = 'unknown';
