+++ assets/js/state/selectors/cluster.js
@@ -1,4 +1,7 @@
-import { getCluster } from '@state/selectors';
+export const getCluster =
+  (id) =>
+  ({ clustersList }) =>
+    clustersList.clusters.find((cluster) => cluster.id === id);
 
 export const getClusterHostIDs =
   (clusterID) =>
@@ -6,6 +9,3 @@ export const getClusterHostIDs =
     hostsList.hosts
       .filter((host) => host.cluster_id === clusterID)
       .map(({ id: hostID }) => hostID);
-
-export const getClusterSelectedChecks = (clusterID) => (state) =>
-  getCluster(clusterID)(state).selected_checks;
